domain/entityのint64をintに

+++ domain/entity/session.go
@@ -3,8 +3,8 @@ package entity
 import "time"
 
 type Session struct {
-	ID			int64
-	UserID 		int64
-	Token		string
-	ExpiresAt	time.Time
+	ID        int
+	UserID    int
+	Token     string
+	ExpiresAt time.Time
 }
